+++ src/main.rs
@@ -1,18 +1,18 @@
 use std::env;
 mod todo; //https://doc.rust-lang.org/book/second-edition/ch07-00-modules.html
 mod tools;
-use crate::tools::directory;
+use crate::todo::*;
 
 fn main() {
     // get arguments
     let args: Vec<String> = env::args().collect();
 
-    let config: directory::Config = directory::get_config();
+    let config: ConfigFile = get_config();
     let handler = todo::TodoHandler::init(&config);
 
     // get the todos
-    let mut parser = todo::TodoParser::new(&config); // get empty parser
-    parser.strings_to_todo(directory::read_lines(&handler.complete_path)); // populate parser
+    let mut parser = todo::Config::new(&config); // get empty parser
+    parser.strings_to_todo(read_lines(&handler.complete_path)); // populate parser
 
     if args.len() > 1 {
         let operation = &args[1];
