+++ src/todo.rs
@@ -1,18 +1,201 @@
 use regex::Regex;
 use std::path::PathBuf;
 
-use crate::tools::directory;
-
 // testing with time crates
 use chrono::prelude::*;
+use dirs;
+use serde::{Deserialize, Serialize};
+use std::fs;
+use std::fs::File;
+use std::io::{self, Write};
+use toml;
 
+enum Error {
+    UserRefused,
+}
+
+pub fn readinput(prompt: &str) -> io::Result<String> {
+    let mut buffer = String::new();
+    print!("{prompt}");
+    io::stdout().flush()?;
+    let stdin = io::stdin();
+    stdin.read_line(&mut buffer)?;
+    // remove trailing newline
+    let input = buffer.trim().to_string();
+
+    Ok(input)
+}
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct ConfigFile {
+    pub format: String, // accepted: "md", "logseq"
+    pub todo_path: String,
+    pub todo_filename: String,
+}
+
+// returns tuple with (directory, filename)
+pub fn ensure_todofile(
+    todo_path: PathBuf,
+    todo_filename: PathBuf,
+) -> Result<(PathBuf, PathBuf), Error> {
+    let unified_path: PathBuf = todo_path.join(todo_filename.clone());
+
+    while !check_for_dir(unified_path.clone()) {
+        match readinput(format!("create {} ? (y/n)", unified_path.display()).as_str())
+            .expect("input failed")
+            .as_str()
+        {
+            "y" => create_path(unified_path.clone()),
+            _ => return Err(Error::UserRefused),
+        }
+    }
+
+    Ok((todo_path, todo_filename))
+}
+
+// should check configuration and if that is invalid, assign arguments as default paths
+pub fn get_config() -> ConfigFile {
+    // defaults for content of config
+    let mut configuration = ConfigFile {
+        path: TodoPath {
+            todo_path: dirs::home_dir().unwrap().to_string_lossy().to_string(),
+            todo_filename: "todo.md".to_string(),
+        },
+        format: "md".to_string()    
+    };
+    // path to the configuration file
+    let path = dirs::config_dir()
+        .expect("config dir error")
+        .join(PathBuf::from("todo-md-rs"))
+        .join(PathBuf::from("config.toml"));
+
+    // check if configuration can be found
+    if check_for_dir(path.clone()) {
+        // config found
+        let contents: String = fs::read_to_string(path).unwrap();
+        configuration = match toml::from_str(&contents) {
+            Ok(content) => content,
+            Err(_e) => {
+                println!("error in your config!");
+
+                println!("using the following defaults: {configuration:?}");
+                configuration
+            }
+        };
+    } else {
+        // config not found
+        match readinput("create base configuraton file? (y/n): ")
+            .expect("input failed")
+            .as_str()
+        {
+            // create file and write defaults into it
+            "y" => {
+                create_path(path.clone());
+
+                let _ = export_line(&path, toml::to_string(&configuration).unwrap());
+            }
+            _ => {
+                println!("using temporary path"); // configuration should not have changed from the initialization
+                let toml = toml::to_string(&configuration).unwrap();
+                println!("{toml:#?}");
+            }
+        }
+    }
+
+    configuration
+}
+
+pub fn read_lines(filepath: &PathBuf) -> Vec<String> {
+    // let parser = crate::todo::TodoParser::new();
+    let mut lines: Vec<String> = Vec::new();
+    let file = fs::read_to_string(filepath).expect("file not found");
+    for line in file.lines() {
+        lines.push(line.to_string())
+    }
+    // for line in file.lines() {
+    //     parser
+    //         .todo_list
+    //         .push(parser.string_to_todo(line.to_string()));
+    //     println!("{line}");
+    // }
+    lines
+}
+
+pub fn remove_lines(filepath: &PathBuf, indices: Vec<u32>) {
+    let original_content = fs::read_to_string(filepath).expect("file not found");
+    let indices_zero_indexed: Vec<u32> = indices
+        .iter()
+        .map(|&nr| if nr < 1 { nr } else { nr - 1 })
+        .collect();
+
+    let lines: Vec<String> = original_content
+        .lines()
+        .enumerate()
+        .filter_map(|(i, line)| {
+            // line number is 1-indexed but we'll remove by counting 0-indexed
+            if !indices_zero_indexed.contains(&(i as u32)) {
+                Some(line.to_string())
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    println!("remove_lines: {:?}", indices);
+
+    let mut file = File::create(filepath).unwrap();
+    for line in lines {
+        let _ = writeln!(file, "{}", line);
+    }
+}
+
+pub fn export_line(filepath: &PathBuf, line: String) -> std::io::Result<()> {
+    let mut file = fs::OpenOptions::new().append(true).open(filepath)?;
+    file.write_all(format!("{line}\n").as_bytes())?;
+
+    let _newone = fs::read_to_string(filepath).unwrap();
+
+    Ok(())
+}
+
+pub fn check_for_dir(path: PathBuf) -> bool {
+    match path.try_exists() {
+        Ok(true) => {
+            // just return true?
+            return true;
+        }
+        Ok(false) => {
+            // just return false?
+            return false;
+        }
+        Err(e) => {
+            // error
+            println!("attention: {e:?}");
+            panic!("unhandled error occured: tried checking for dir");
+        }
+    }
+}
+
+pub fn create_path(path: PathBuf) {
+    match check_for_dir(path.clone()) {
+        false => {
+            if let Some(parent) = path.parent() {
+                let _ = std::fs::create_dir_all(parent);
+            }
+            println!("creating: {path:?}");
+            let _ = File::create(path.clone()).expect("creating file failed");
+        }
+        true => {
+            println!("file already exists");
+        }
+    }
+}
 fn get_date() -> String {
     let date = Local::now();
     date.to_string()
 }
 
 #[derive(Debug, Clone)]
-pub struct Todo {
+pub struct TodoItem {
     pub id: u32,
     pub line: u32, // should probably be a tuple of start/end (see logseq task block would remain junk)
     pub is_completed: bool,
@@ -20,9 +203,9 @@ pub struct Todo {
     pub creation_date: String,
     pub title: String,
 }
-impl Todo {
-    pub fn new() -> Todo {
-        Todo {
+impl TodoItem {
+    pub fn new() -> TodoItem {
+        TodoItem {
             id: 0,
             line: 0,
             is_completed: false,
@@ -32,7 +215,7 @@ impl Todo {
         }
     }
 
-    pub fn get_string(todoitem: Todo, parser: TodoParser) -> String {
+    pub fn get_string(todoitem: TodoItem, parser: Config) -> String {
         let mut result_string = String::new();
         result_string.push_str(if todoitem.is_completed {
             &parser.example_done
@@ -51,15 +234,30 @@ impl Todo {
     }
 }
 
-#[derive(Debug)]
-pub struct TodoHandler {
+// TODO: not respecting config yet
+pub struct Config {
+    // from TodoParser
+    pub completion_style: Regex, // check if line is valid
+    pub completion_done: Regex,  // check if valid line is done
+    pub example_todo: String,
+    pub example_done: String,
+    pub todo_list: Vec<TodoItem>,
+    // from TodoHandler
     pub path: PathBuf,
     pub filename: PathBuf,
     pub complete_path: PathBuf,
 }
-impl TodoHandler {
-    pub fn init(config: &directory::Config) -> TodoHandler {
-        match directory::ensure_todofile(
+impl Config {
+
+    //
+    // converted
+    //
+
+    //
+    // from TodoHandler
+    //
+    pub fn init(config: &ConfigFile) -> TodoHandler {
+        match ensure_todofile(
             config.path.todo_path.clone().into(),
             config.path.todo_filename.clone().into(),
         ) {
@@ -77,20 +275,20 @@ impl TodoHandler {
         }
     }
 
-    pub fn add(&self, input_content: Vec<String>, parser: TodoParser) {
-        let mut todoitem: Todo = Todo::new();
+    pub fn add(&self, input_content: Vec<String>, parser: Config) {
+        let mut todoitem: TodoItem = TodoItem::new();
 
         let title = input_content.join(" ").trim().to_string();
         println!("adding task: {}", title);
         todoitem.title = title;
         // write item into file
-        let _ = directory::export_line(
+        let _ = export_line(
             &self.path.join(&self.filename),
-            Todo::get_string(todoitem, parser),
+            TodoItem::get_string(todoitem, parser),
         );
     }
 
-    pub fn done(&self, indicies: Vec<String>, mut parser: TodoParser) {
+    pub fn done(&self, indicies: Vec<String>, mut parser: Config) {
 
         // keep track of task_ids to then mark as done
         let mut to_check_off: Vec<usize> = Vec::new();
@@ -121,7 +319,7 @@ impl TodoHandler {
     }
 
     // TODO notice that we're currently removing by line number not by task id
-    pub fn remove(&self, id: Vec<String>, path: PathBuf, parser: TodoParser) {
+    pub fn remove(&self, id: Vec<String>, path: PathBuf, parser: Config) {
         // sanitize the given arguments
         let mut sanitized_ids: Vec<u32> = Vec::new();
         for item in id {
@@ -139,10 +337,10 @@ impl TodoHandler {
                 lines_to_rm.push(item.line)
             }
         }
-        directory::remove_lines(&path, lines_to_rm);
+        remove_lines(&path, lines_to_rm);
     }
 
-    pub fn list(&self, todos: Vec<Todo>) {
+    pub fn list(&self, todos: Vec<TodoItem>) {
         for todoitem in todos {
             if todoitem.is_completed {
                 println!("[X] {} {}", todoitem.id, todoitem.title);
@@ -151,19 +349,12 @@ impl TodoHandler {
             }
         }
     }
-}
 
-// TODO: not respecting config yet
-pub struct TodoParser {
-    pub completion_style: Regex, // check if line is valid
-    pub completion_done: Regex,  // check if valid line is done
-    pub example_todo: String,
-    pub example_done: String,
-    pub todo_list: Vec<Todo>,
-}
-impl TodoParser {
-    pub fn new(config: &directory::Config) -> TodoParser {
-        let default_md: TodoParser = TodoParser {
+    //
+    // from TodoParser
+    //
+    pub fn new(config: &ConfigFile) -> Config {
+        let default_md: Config = Config {
             completion_style: Regex::new(r"^\s*-\s*\[[ xX]\]").unwrap(),
             completion_done: Regex::new(r"^\s*-\s*\[[^\s]\]").unwrap(),
             example_todo: String::from("- [ ]"),
@@ -172,12 +363,12 @@ impl TodoParser {
         };
 
         // Markdown style
-        if config.format.checkbox_style == "md" {
+        if config.format == "md" {
             return default_md;
         }
         // Logseq style
-        else if config.format.checkbox_style == "logseq" {
-            return TodoParser {
+        else if config.format == "logseq" {
+            return Config {
                 completion_style: Regex::new(r"^\s*-\s*[A-Z]{4}").unwrap(),
                 completion_done: Regex::new(r"^\s*-\s*DONE\s").unwrap(),
                 example_todo: String::from("- TODO"),
@@ -193,12 +384,12 @@ impl TodoParser {
     }
     // TODO: complete this method to include all the other fields of Todo
     pub fn strings_to_todo(&mut self, lines: Vec<String>) {
-        let mut item_list: Vec<Todo> = Vec::new();
+        let mut item_list: Vec<TodoItem> = Vec::new();
 
         for (linecount, line) in lines.iter().enumerate() {
             // new task detected
             if self.completion_style.is_match(&line) {
-                let mut item = Todo::new();
+                let mut item = TodoItem::new();
 
                 item.id = item_list.len() as u32 + 1;
