+++ src/tools/error.rs
@@ -1,4 +1,1 @@
-pub enum DirErrors {
-    //CheckingFileFailed,
-    UserRefused,
-}
+
